app: use absolute offset for filter-tool split-preview guide position
In GimpDrawableFilter and GimpFilterTool, use an absolute offset
for the split-preview guide position, instead of storing it as a
fraction of the drawable's width/height.  The latter could
introduce rounding error, which would result in wrong coordinates
when converted back to an absolute offset.

(cherry picked from commit 214936e8c6d14e17249d06d1b02aaf53b18f73b2)

+++ app/core/gimpdrawablefilter.h
@@ -72,7 +72,7 @@ void       gimp_drawable_filter_set_preview_split
                                                (GimpDrawableFilter  *filter,
                                                 gboolean             enabled,
                                                 GimpAlignmentType    alignment,
-                                                gdouble              split_position);
+                                                gint                 split_position);
 void       gimp_drawable_filter_set_opacity    (GimpDrawableFilter  *filter,
                                                 gdouble              opacity);
 void       gimp_drawable_filter_set_mode       (GimpDrawableFilter  *filter,
